Expressions now return parse trees!
parsimonious has gone from being merely a recognizer to being a parser! We now cache entire subtrees rather than just lengths, because otherwise we'd have to continually reconstitute nodes out of the cached ints. This is probably horrible for RAM use, but we'll get to that later. At least we're using __slots__.