Make our first level-1 integration test pass, mostly by adding $ to the eol pattern.
The level-0 expression tree is used to parse the DSL grammar which is itself written in the DSL. The result is the level-1 expression tree, which we test here. We should then reparse the written-in-DSL DSL grammar with the level-1 tree, either as part of the bootstrap at import time or, at least, in a test, to ensure equivalence.