Significantly improve reprs of Expressions and Grammars. Grammar reprs are round-trippable.
Now it's clear not only what an Expression's class is but also what its entire tree is, in terms of the same rule syntax that created it.

Factor up the thing where __str__ is just the encoded __unicode__. Not sure I like this.

Add test for #8: test_unicode_crash. This should have been part of f19557bc6440697b394368c48331dc2af0894676.