replace getargspec with  getfullargspec
`getargspec()` is deprecated. `getfullargspec()` is supposed to be a drop in replacement for most use cases, expanding the return value of `getargspec()` to include function annotations and keyword-only parameters. The exact behavior changed in 3.6 and again in 3.7, so if this becomes a problem look into directly calling `signature()`, which underlies `getfullargspec()`, in order to build exactly what is needed.