re-adding the unicode flag to the documentation about regex
The regex lib has more possible flags than the base re module. One of these restores the python 2.x re modules unicode flag, which forces the regex lib to treat the string like Unicode. Without it, it will test the type first and evaluate it accordingly. Parsimonious tests for handling unicode via a flag already, so it is useful to restore this in the documentation.

We may consider adding support for the other flags supported by regex in the future